fixed compiled warning
Former-commit-id: fe85d9b0680bf8dc1b9e0d5004a3cd8f724b716f

+++ src/rpc/rpctx.cpp
@@ -1562,17 +1562,17 @@ typedef struct {
 	  IMPLEMENT_SERIALIZE
 	    (
 	        READWRITE(status);
-	  for(int i = 0 ;i < sizeof(sendid);i++)
+	  for(size_t i = 0 ;i < sizeof(sendid);i++)
 	        READWRITE(sendid[i]);
-	  for(int i = 0 ;i < sizeof(acceptid);i++)
+	  for(size_t i = 0 ;i < sizeof(acceptid);i++)
 	        READWRITE(acceptid[i]);
 	  READWRITE(money);
 	  READWRITE(hight);
 	  READWRITE(delyhight);
 	  READWRITE(shash);
-	  for(int i = 0 ;i < sizeof(sdata);i++)
+	  for(size_t i = 0 ;i < sizeof(sdata);i++)
 		        READWRITE(sdata[i]);
-	  for(int i = 0 ;i < sizeof(adata);i++)
+	  for(size_t i = 0 ;i < sizeof(adata);i++)
 		        READWRITE(adata[i]);
 	    )
